Begin moving components to test project.
